tests/e2e/requirements.txt
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
pytest-timeout==2.3.1
herd-client
websockets==12.0
